+++ Carbon/Functions/Clear-DscLocalResourceCache.ps1
@@ -21,21 +21,9 @@ function Clear-CDscLocalResourceCache
 
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    if( $PSVersionTable['PSEdition'] -eq 'Core' )
-    {
-        Get-CimInstance msft_providers | 
-        Where-Object {$_.provider -like 'dsccore'} | 
-        Select-Object -ExpandProperty HostProcessIdentifier | 
-        ForEach-Object { Get-Process -ID $_ } | 
-        Stop-Process -Force
-    }
-    else
-    {
-        Get-WmiObject msft_providers | 
-        Where-Object {$_.provider -like 'dsccore'} | 
-        Select-Object -ExpandProperty HostProcessIdentifier | 
-        ForEach-Object { Get-Process -ID $_ } | 
-        Stop-Process -Force
-    }
-
+    Get-Cim -Class 'msft_providers' | 
+    Where-Object {$_.provider -like 'dsccore'} | 
+    Select-Object -ExpandProperty HostProcessIdentifier | 
+    ForEach-Object { Get-Process -ID $_ } | 
+    Stop-Process -Force
 }