+++ Carbon/Functions/Get-FileSharePermission.ps1
@@ -78,13 +78,21 @@ function Get-CFileSharePermission
     }
         
     $acl = $null  
-    $lsss = Get-WmiObject -Class 'Win32_LogicalShareSecuritySetting' -Filter "name='$Name'"
+    $lsss = Get-Cim -Class 'Win32_LogicalShareSecuritySetting' -Filter "name='$Name'"
     if( -not $lsss )
     {
         return
     }
 
-    $result = $lsss.GetSecurityDescriptor()
+    if( $PSVersionTable.PSEdition -eq 'Core' )
+    {
+        $result = Invoke-CimMethod -InputObject $lsss -MethodName 'GetSecurityDescriptor'
+    }
+    else
+    {
+        $result = $lsss.GetSecurityDescriptor()
+    }
+
     if( -not $result )
     {
         return