+++ Carbon/Functions/Get-WmiLocalUserAccount.ps1
@@ -31,6 +31,6 @@ function Get-CWmiLocalUserAccount
 
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
     
-    return Get-WmiObject Win32_UserAccount -Filter "Domain='$($env:ComputerName)' and Name='$Username'"
+    return Get-Cim -Class 'Win32_UserAccount' -Filter "Domain='$($env:ComputerName)' and Name='$Username'"
 }