+++ Test/MSMQ/Test-InstallMSMQMessageQueue.ps1
@@ -12,7 +12,7 @@
 
 # These tests should only run if MSMQ is not installed
 $userDomain = $env:USERDNSDOMAIN
-$computerDomain = Get-WmiObject 'Win32_ComputerSystem' | Select-Object -ExpandProperty Domain
+$computerDomain = Get-Cim -Class 'Win32_ComputerSystem' | Select-Object -ExpandProperty Domain
 if( (Get-Service -Name MSMQ -ErrorAction SilentlyContinue) -and $userDomain -eq $computerDomain )
 {