+++ Test/UsersAndGroups/Test-GetWmiLocalUserAccount.ps1
@@ -17,7 +17,7 @@ function Start-TestFixture
 
 function Test-ShouldGetUser
 {
-    Get-WmiObject Win32_UserAccount -Filter "Domain='$($env:ComputerName)'" | ForEach-Object {
+    Get-Cim -Class 'Win32_UserAccount' -Filter "Domain='$($env:ComputerName)'" | ForEach-Object {
         $user = Get-WmiLocalUserAccount -Username $_.Name
         Assert-NotNull $user
         Assert-Equal $_.Name $user.Name